api compatiblity fix
